Add files via upload
Algorithm visualizer helps to visualise the sorting algorithm in graphical form . Technology used to make this project is HTML,CSS, and Java Script. For making algorithm visualizer most of part done by using Java Script .
Features of Java Script in   algorithm visualizer are - increase or decrease the size of bar, Increase or Decrease the speed of bar , with the help of different - different function we decide the movement of bar according to different - different sorting algorithm like Bubble sort , Insertion Sort , Merge Sort with the help of HTML and Java script we also make a Drop down button where we Change the Algorithm according to the choice. So this project is very helpful to understand the concept of Algorithm and its functionality